+++ data_pipelines_cli/cli_commands/deploy.py
@@ -5,6 +5,7 @@ from typing import Any, Dict, Optional, cast
 import click
 import yaml
 
+from ..bi_utils import BiAction, bi
 from ..cli_configs import find_datahub_config_file
 from ..cli_constants import BUILD_DIR
 from ..cli_utils import echo_error, echo_info, subprocess_run
@@ -35,6 +36,7 @@ class DeployCommand:
     """Dictionary of arguments required by a specific cloud storage provider,
     e.g. path to a token, username, password, etc."""
     env: str
+    bi_git_key_path: str
 
     def __init__(
         self,
@@ -43,11 +45,13 @@ class DeployCommand:
         dags_path: Optional[str],
         provider_kwargs_dict: Optional[Dict[str, Any]],
         datahub_ingest: bool,
+        bi_git_key_path: str,
     ) -> None:
         self.docker_args = DockerArgs(env, None, {}) if docker_push else None
         self.datahub_ingest = datahub_ingest
         self.provider_kwargs_dict = provider_kwargs_dict or {}
         self.env = env
+        self.bi_git_key_path = bi_git_key_path
 
         try:
             self.blob_address_path = (
@@ -73,7 +77,12 @@ class DeployCommand:
         if self.datahub_ingest:
             self._datahub_ingest()
 
+        self._bi_push()
+
         self._sync_bucket()
+
+    def _bi_push(self) -> None:
+        bi(self.env, BiAction.DEPLOY, self.bi_git_key_path)
 
     def _docker_push(self) -> None:
         """
@@ -156,12 +165,19 @@ class DeployCommand:
     default=False,
     help="Whether to ingest DataHub metadata",
 )
+@click.option(
+    "--bi-git-key-path",
+    type=str,
+    required=False,
+    help="Path to the key with write access to repo",
+)
 def deploy_command(
     env: str,
     dags_path: Optional[str],
     blob_args: Optional[io.TextIOWrapper],
     docker_push: bool,
     datahub_ingest: bool,
+    bi_git_key_path: str,
 ) -> None:
     if blob_args:
         try:
@@ -173,9 +189,5 @@ def deploy_command(
         provider_kwargs_dict = None
 
     DeployCommand(
-        env,
-        docker_push,
-        dags_path,
-        provider_kwargs_dict,
-        datahub_ingest,
+        env, docker_push, dags_path, provider_kwargs_dict, datahub_ingest, bi_git_key_path
     ).deploy()