+++ tests/cli_commands/test_deploy.py
@@ -85,6 +85,7 @@ class DeployCommandTestCase(unittest.TestCase):
                     _blob_address,
                     provider_kwargs_dict,
                     _datahub_ingest,
+                    _bi_git_key_path,
                 ):
                     nonlocal result_provider_kwargs
                     result_provider_kwargs = provider_kwargs_dict
@@ -106,7 +107,9 @@ class DeployCommandTestCase(unittest.TestCase):
         # 'filesystem_utils.LocalRemoteSync' here. It gets tested in
         # a dedicated 'test_filesystem_utils' file.
         runner = CliRunner()
-        with patch("pathlib.Path.cwd", lambda: self.dbt_project_config_dir):
+        with patch("pathlib.Path.cwd", lambda: self.dbt_project_config_dir), patch(
+            "data_pipelines_cli.cli_commands.deploy.bi"
+        ):
             result = runner.invoke(
                 _cli,
                 [
@@ -152,15 +155,19 @@ class DeployCommandTestCase(unittest.TestCase):
             "pathlib.Path.cwd", lambda: self.dbt_project_config_dir
         ):
             with self.assertRaises(DependencyNotInstalledError):
-                DeployCommand("base", False, self.storage_uri, self.provider_args, True).deploy()
+                DeployCommand(
+                    "base", False, self.storage_uri, self.provider_args, True, None
+                ).deploy()
 
     @patch("data_pipelines_cli.cli_commands.deploy.BUILD_DIR", goldens_dir_path)
     @patch("data_pipelines_cli.cli_configs.BUILD_DIR", goldens_dir_path)
     def test_datahub_run(self):
         with patch("pathlib.Path.cwd", lambda: self.dbt_project_config_dir), patch(
             "data_pipelines_cli.cli_commands.deploy.subprocess_run", self._mock_run
-        ), patch.dict("sys.modules", datahub=MagicMock()):
-            DeployCommand("base", False, self.storage_uri, self.provider_args, True).deploy()
+        ), patch.dict("sys.modules", datahub=MagicMock()), patch(
+            "data_pipelines_cli.cli_commands.deploy.bi"
+        ):
+            DeployCommand("base", False, self.storage_uri, self.provider_args, True, None).deploy()
             self.assertListEqual(
                 [
                     "datahub",
@@ -176,7 +183,9 @@ class DeployCommandTestCase(unittest.TestCase):
             "pathlib.Path.cwd", lambda: self.dbt_project_config_dir
         ), patch("data_pipelines_cli.cli_constants.BUILD_DIR", self.build_temp_dir):
             with self.assertRaises(DependencyNotInstalledError):
-                DeployCommand("base", True, self.storage_uri, self.provider_args, False).deploy()
+                DeployCommand(
+                    "base", True, self.storage_uri, self.provider_args, False, None
+                ).deploy()
 
     @patch(
         "data_pipelines_cli.cli_commands.deploy.BUILD_DIR",
@@ -184,7 +193,7 @@ class DeployCommandTestCase(unittest.TestCase):
     )
     def test_no_airflow_address(self):
         with self.assertRaises(AirflowDagsPathKeyError):
-            DeployCommand("base", False, None, None, False)
+            DeployCommand("base", False, None, None, False, None)
 
     def test_airflow_address(self):
         with tempfile.TemporaryDirectory() as tmp_dir, patch(
@@ -199,7 +208,7 @@ class DeployCommandTestCase(unittest.TestCase):
                 tmp_airflow_path,
             )
 
-            deploy_command = DeployCommand("base", False, None, None, False)
+            deploy_command = DeployCommand("base", False, None, None, False, None)
         self.assertEqual(
             "gcs://test-sync-project/sync-dir/dags/my-project-name",
             deploy_command.blob_address_path,
@@ -221,7 +230,7 @@ class DeployCommandTestCase(unittest.TestCase):
                     tmp_file_path,
                 )
 
-            deploy_command = DeployCommand("staging", False, None, None, False)
+            deploy_command = DeployCommand("staging", False, None, None, False, None)
         self.assertEqual(
             "gcs://test/jinja/path/com/my/project/name",
             deploy_command.blob_address_path,
@@ -249,8 +258,10 @@ class DeployCommandTestCase(unittest.TestCase):
             "data_pipelines_cli.data_structures.git_revision_hash", lambda: "sha1234"
         ), patch(
             "data_pipelines_cli.cli_constants.BUILD_DIR", self.build_temp_dir
+        ), patch(
+            "data_pipelines_cli.cli_commands.deploy.bi"
         ):
-            DeployCommand("base", True, self.storage_uri, self.provider_args, False).deploy()
+            DeployCommand("base", True, self.storage_uri, self.provider_args, False, None).deploy()
 
         self.assertEqual("my_docker_repository_uri", docker_kwargs.get("repository"))
         self.assertEqual("sha1234", docker_kwargs.get("tag"))
@@ -278,4 +289,6 @@ class DeployCommandTestCase(unittest.TestCase):
             "data_pipelines_cli.cli_constants.BUILD_DIR", self.build_temp_dir
         ):
             with self.assertRaises(DataPipelinesError):
-                DeployCommand("base", True, self.storage_uri, self.provider_args, False).deploy()
+                DeployCommand(
+                    "base", True, self.storage_uri, self.provider_args, False, None
+                ).deploy()